App2/requirements.txt
streamlit>=1.20.0
scikit-image>=0.21.0
numpy>=1.22.0
opencv-python-headless>=4.8.0
Pillow>=9.0.0
